Separate validations from register user mutation
